requirements.txt
llama-index-llms-azure-openai==0.1.8
Jinja2==3.1.6
PyPDF2==3.0.1
pytesseract==0.3.13
pdf2image==1.17.0